backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Telegram
telethon==1.34.0
cryptg==0.4.0

# AI & ML
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2
sentence-transformers==2.3.1
numpy==1.26.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
aiofiles==23.2.1
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
